aider: Move the `publicPath` option inside a `devMiddleware` property in webpack dev server configuration.
# Aider chat conversation:

USER: @BenGWeeks ➜ /workspaces/AzureDevOpsExtensions (main) $ npm start

> azure-devops-helloworld-widget@1.0.0 start
> webpack-dev-server --open

[webpack-cli] Invalid options object. Dev Server has been initialized using an options object that does not match the API schema.
 - options has an unknown property 'publicPath'. These properties are valid:
   object { allowedHosts?, bonjour?, client?, compress?, devMiddleware?, headers?, historyApiFallback?, host?, hot?, http2?, https?, ipc?, liveReload?, magicHtml?, onAfterSetupMiddleware?, onBeforeSetupMiddleware?, onListening?, open?, port?, proxy?, server?, setupExitSignals?, setupMiddlewares?, static?, watchFiles?, webSocketServer? }
@BenGWeeks ➜ /workspaces/AzureDevOpsExtensions (main) $

ASSISTANT: The error message indicates that the `publicPath` property is not recognized in the options object for the webpack dev server. This is likely due to changes in the API schema for the webpack dev server.

In the latest versions of webpack-dev-server, the `publicPath` option should be inside the `devMiddleware` property.

Let's modify your `webpack.config.js` file to reflect this change:

webpack.config.js
```javascript
<<<<<<< SEARCH
devServer: {
    contentBase: path.join(__dirname, '/'),
    publicPath: '/',
    compress: true,
    port: 9000
},
=======
devServer: {
    static: {
        directory: path.join(__dirname, '/'),
    },
    devMiddleware: {
        publicPath: '/',
    },
    compress: true,
    port: 9000
},
>>>>>>> REPLACE
```

This change moves the `publicPath` option inside a `devMiddleware` property, which should resolve the error. After making this change, try running `npm start` again and see if the issue is resolved.